deduplicate ethDB logic
